Add Presentation Model
The `Presentation` model is pretty straight forward. Setting up the
associations to `Event` and `Person`.

Validations for `Presentation` are also pretty relaxed. Be aware that
`presenter` and `person` are conditionally dependent on eachother. If
one is set, the other cannot be.

`Presentation` adds a new property of `presenter_name` which will return
either the `presenter` if present, or the association
`Person`.`first_name`

The Presentation Factory includes associations for `Person` and `Event`
for ease of testing. If you need to test `presenter` you will need to
unset person or the model will be invalid.