Clean Up Relations
Removing legacy has_many_through association on `Person`. Due to the
nature of the new `Presentation` we no longer have a `Presentation`
with the potential for multiple presenters.

Once the legacy has_many_through associations area removed, we need to
add back as has_many association back to `Person`. At the same time, we
will add a has_many association to `Event`.