Migration - Alter Presentations
The purpose of this is to update and clean up the legacy presentations
table in the database, and update the factories to match.

Major change was to remove the date column; Presentation will now be
part of an event which controls the date.

Minor changes were renaming of columns `title` => `topic` and `content`
=> `description`. As well, adding `presenter` for non-people presenters,
and `duration`.

Adding in references to `Event` and `People` with indexes for handling
the updated associations

There are still some association for `Presentations` that will be
cleaned up in the next commit dealing with the model
implementation of presentations. It would not be advisable to cherry
pick this commit and leave out the following ones.