Refactor spec to work as intended
This spec was passing but not quite doing what we intended it to. I've made some
refactors to bring it back to where it should be.

Replaced some .length calls with static integers just so we can be a little bit
more clear with what exactly is going on here.