Update presentations policy
We are not going to be limiting presentation creation to
admin users anymore. This commit sets the presentation policy
to reflect that.

I've also removed the admin context from the presentation
controller specs. The reason for this being is there are
no actions in the controller as of now in which an admin
has higher priveleges over a regular user.