Refactor Pundit Policies
I've refactored the pundit policies to all always
be operating on a person object. Because of this change
I also wrote a signed_in? method to improve readability.

Knowing we are always operating on a Person object in our
policies let's us clean up our code a bit and not have to consistenly
check for nil / safe navigation.